refactor: create more comp directories
