+++ src/app/(public)/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from '@/components/LoginForm';
+import { LoginForm } from '@/components/form/LoginForm';
 
 export default function Page() {
   return <LoginForm />;
